refactor docker/tools
