+++ src/docker/tools/index.go
@@ -1,4 +1,4 @@
-package docker
+package tools
 
 import (
 	"context"
